requirements.txt
pygame-ce
requests
voluptuous
pyyaml
jq

